requirements.txt
matplotlib==3.0.3
numpy==1.16.3
pandas==0.24.2
Pillow==6.0.0
plotly==3.8.1
wordcloud==1.5.0
